fix: Lv3. 비밀번호가 일치하는 경우만 유저, 일정 수정 삭제 가능하도록 수정

+++ src/main/java/com/example/schedule/service/ScheduleService.java
@@ -1,6 +1,8 @@
 package com.example.schedule.service;
 
+import com.example.schedule.dto.request.DeleteRequestDto;
 import com.example.schedule.dto.request.ScheduleRequestDto;
+import com.example.schedule.dto.request.UpdateScheduleRequestDto;
 import com.example.schedule.dto.response.ScheduleResponseDto;
 import com.example.schedule.entity.Schedule;
 import com.example.schedule.entity.User;
@@ -8,8 +10,10 @@ import com.example.schedule.repository.ScheduleRepository;
 import com.example.schedule.repository.UserRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
@@ -21,6 +25,10 @@ public class ScheduleService {
     public ScheduleResponseDto save(ScheduleRequestDto dto) {
 
         User findUser = userRepository.findByUserName(dto.getUserName());
+
+        if (!findUser.getPassword().equals(dto.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "비밀번호가 일치하지 않습니다.");
+        }
 
         Schedule schedule = new Schedule(findUser, dto.getTitle(), dto.getContents());
         Schedule savedSchedule = scheduleRepository.save(schedule);
@@ -46,17 +54,26 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResponseDto updateSchedule(Long id, ScheduleRequestDto dto) {
+    public ScheduleResponseDto updateSchedule(Long id, UpdateScheduleRequestDto dto) {
 
         Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+
+        if (!findSchedule.getUser().getPassword().equals(dto.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "비밀번호가 일치하지 않습니다.");
+        }
+
         findSchedule.updateSchedule(dto.getTitle(), dto.getContents());
 
         return ScheduleResponseDto.toDto(findSchedule);
     }
 
-    public void delete(Long id) {
+    public void delete(Long id, DeleteRequestDto dto) {
 
         Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+
+        if (!findSchedule.getUser().getPassword().equals(dto.getPassword())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "비밀번호가 일치하지 않습니다.");
+        }
 
         scheduleRepository.delete(findSchedule);
     }
